Fix pylint.lint.Run initialization error
When Pylava is run with Pylava_pylint (fork of Pylama_pylint), the
following error occurs:

    $ pylava -l pylint
    Traceback (most recent call last):
      File "venv/pylava/lib/python3.7/site-packages/pylava/core.py",
      line 65, in run
        select=params.get("select", set()), params=lparams)
      File "venv/pylava/lib/python3.7/site-packages/
      pylava_pylint-0.0.0-py3.7.egg/pylava_pylint/main.py",
      line 57, in run
        Run([path] + params.to_attrs(), reporter=reporter, exit=False)
    TypeError: __init__() got an unexpected keyword argument 'exit'

This issue is present in Pylama_pylint which is inherited by
Pylava_pylint as a result of it being a fork of the former.

The `Run` class in the error message above belongs to the `pylint.lint`
module. The initialization of this `Run` class fails because since
Pylint version 2.0, the `__init__()` method of this class no longer has
a keyword argument named `exit`. The keyword argument is `do_exit()`
instead.

In [commit 4210ef9][1] of Pylint (released in Pylint 2.0), the `exit`
parameter was renamed to `do_exit` parameter which causes Pylama_pylint
and its fork Pylava_pylint to break.

This commit fixes this issue by initializing the `Run` instance with the
correct keyword parameter.

[1]: https://github.com/PyCQA/pylint/commit/4210ef9